change values from int to long, more work on request mapping

+++ src/main/java/com/coinportfolio/server/models/Coin.java
@@ -37,7 +37,7 @@ public class Coin {
         currencyValues.put(currencyTicker, rate);
     }
 
-    public int getValue(CurrenciesEnum currencyTicker){
+    public long getValue(CurrenciesEnum currencyTicker){
         return currencyValues.get(currencyTicker).getValue();
     }
